Rename publication plots to publication benchmark plots
